Move CM5-only utilities into the v5/ directory
